Faulty BNO085 footprint
First, I do not like you UltraLibrarian anymore, for you have wasted a
lot of time and effort (and money) on footprints that are just straight
up wrong. I am this close to unleashing unbridled rage on this but will
bottle it up for now.

Second, I took the BNO085 footprint from the official SlimeVR tracker
and put it in this project. At least they have the right footprint...

Stencils for the main microcontroller does not need to change but the
extension module got a revamp in favor of being more versatile. You can
now choose which I2C address the sensor operates on; just make sure you
bridge the right contacts.